eLibrary/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
drf_spectacular==0.24.2
